requirements.txt
ccxt>=4.4.0
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0
python-dotenv>=1.0.0
rich>=13.7.0
